requirements.txt
Flask==3.0.3
Pillow==11.0.0

